+++ frontend/src/pages/Admin/Course/Add Course/Course Content/SubSectionFields.jsx
@@ -2,7 +2,7 @@ import { FaPlus } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
 
 import { useFieldArray } from "react-hook-form";
-import InputField from "../../Input Field";
+import InputField from "../../../../../components/Input Field/index";
 
 export default function SubSectionFields({ control, sectionIndex, errors }) {
   const {
@@ -31,6 +31,7 @@ export default function SubSectionFields({ control, sectionIndex, errors }) {
                   )}
                   placeholder="Subsection Title"
                 />
+                
 
                 {Array.isArray(errors?.sections) &&
                   errors?.sections[sectionIndex]?.subsections[subIndex]
