+++ frontend/src/pages/Admin/Course/Add Course/Schema/secondStepValidationSchema.js
