+++ frontend/src/pages/Admin/Course/Add Course/Schema/thirdStepValidationSchema.js
